#15 Combined parser and lexer errors + fixed test script
